기존에 컨테이너 안에 있던 박스를 컴포넌트화하여 다른 파일로 분리
