+++ src/components/member/searchedItems/Item.js
@@ -1,0 +1,23 @@
+import { Link } from "react-router-dom";
+
+const Item = ({ idx, className }) => {
+  return (
+    <Link to={`/jobs/${idx}`} key={idx} className="block w-full h-80">
+      <div className="flex justify-center items-center w-full h-3/5 rounded-md overflow-hidden shadow-md">
+        <img
+          src="https://www.cvpartners.co.kr/Content/images/house/1/5.png"
+          alt="company img"
+          className="w-full h-full object-cover"
+        />
+      </div>
+      <div className="w-full h-2/5 relative">
+        <p className="py-2 text-lg font-bold">업종</p>
+        <p className="text-neutral-500">기업명</p>
+        <p className="py-2 text-neutral-500">주소</p>
+        <p className="mt-5 font-semibold absolute bottom-2">임금 (원)</p>
+      </div>
+    </Link>
+  );
+};
+
+export default Item;
